Refactor: Modularize JavaScript codebase
- Decomposed the monolithic js/script.js into separate modules for state, UI, API, and charts.
- Replaced the old script loading in index.html with a single module-based entry point.
- Removed the now-obsolete js/data.js and js/script.js files.
