+++ index.html
@@ -119,7 +119,6 @@
     <div id="modal-popup" class="modal-popup">
         <div id="modal-content"></div>
     </div>
-    <script src="js/data.js"></script>
-    <script src="js/script.js"></script>
+    <script type="module" src="js/main.js"></script>
 </body>
 </html>
